+++ server/server.ts
@@ -1,11 +1,14 @@
 import * as express from "express";
 import * as http from "http";
 import * as next from "next";
-import * as sockIO from "socket.io";
+import { wsJogoDaVelha } from "./WebSocket/jogodavelha";
+import { wsMensagens } from "./WebSocket/mensagens";
 
 const app = express();
 const httpServer = http.createServer(app);
-const io = sockIO(httpServer);
+
+wsJogoDaVelha(httpServer);
+wsMensagens(httpServer);
 
 const porta = process.env.PORT || 3000;
 // const porta = process.env.PORT || 3000;
@@ -13,23 +16,6 @@ const dev = process.env.NODE_ENV !== "production";
 
 const nextAPP = next({ dev, dir: "./dist" });
 const nextHandler = nextAPP.getRequestHandler();
-
-io.on("connect", (client) => {
-  // tslint:disable-next-line:no-console
-  console.log("CONECTOU PATRÃO");
-
-  // tslint:disable-next-line:no-console
-  client.on("enviaMSG", (msg) => io.emit("recebeMSG", msg));
-
-  client.on("fezJogada", (jogada, vezJogador) =>
-    client.broadcast.emit("fezJogada", jogada, vezJogador)
-  );
-
-  client.on("disconnect", () => {
-    // tslint:disable-next-line:no-console
-    console.log("DISCONECTOU");
-  });
-});
 
 nextAPP.prepare().then(() => {
   app.get("/", (req, res) => {
